requirements_build.txt
# 构建exe所需的依赖包
Pillow>=9.0.0
pyinstaller>=5.0.0